+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## UNRELEASED
+## 1.2.4 - 2026-07-30
 - Looting an Item Frame is now a crime.
 - Picking up entities with CarryOn mod is now a crime.